Split some components
