+++ src/App.js
@@ -1,45 +1,11 @@
-import React, { Component } from 'react';
-import findDogs from './dogsService';
+import React from 'react';
 import Provider from './Provider';
 import Layout from './Layout';
 
-class App extends Component {
-	state = {
-		filters: {
-			size: 3,
-			energy: 3,
-			barkness: 3,
-		},
-		dogs: [],
-	};
-
-	componentDidMount() {
-		this.fetchDogs();
-	}
-
-	fetchDogs = () => {
-		findDogs(this.state.filters).then(dogs => this.setState({ dogs }));
-	};
-
-	handleFilterUpdate = (field, value) => {
-		this.setState(
-			prevState => ({
-				filters: {
-					...prevState.filters,
-					[field]: Number(value),
-				},
-			}),
-			this.fetchDogs
-		);
-	};
-
-	render() {
-		return (
-			<Provider>
-				<Layout {...this.state} onFilterUpdate={this.handleFilterUpdate} />
-			</Provider>
-		);
-	}
-}
+const App = () => (
+	<Provider>
+		<Layout />
+	</Provider>
+);
 
 export default App;
