+++ src/Layout.js
@@ -1,19 +1,11 @@
 import React from 'react';
-import { Flex, Box } from 'grid-styled';
+import { Box } from 'grid-styled';
 import Card from './components/Card';
-import DogPhoto from './components/DogPhoto';
 import Footer from './components/Footer';
 import Heading from './components/Heading';
 import Page from './components/Page';
-import Slider from './components/Slider';
+import DogFinder from './components/DogFinderContainer';
 import SubscriptionForm from './components/SubscriptionFormContainer';
-import Widget from './components/Widget';
-
-const SliderColumn = ({ children }) => (
-	<Box w={[1, 1 / 3]} px={2} py={[2, 0]}>
-		{children}
-	</Box>
-);
 
 const Layout = ({ filters, dogs, onFilterUpdate }) => (
 	<Page.Container>
@@ -25,43 +17,7 @@ const Layout = ({ filters, dogs, onFilterUpdate }) => (
 				</span>
 			</Heading>
 			<Box mb={4}>
-				<Card mb={3}>
-					<Flex mx={-2} my={[-2, 0]} flexWrap="wrap">
-						<SliderColumn>
-							<Widget label="Size">
-								<Slider
-									value={filters.size}
-									onChange={x => onFilterUpdate('size', x)}
-								/>
-							</Widget>
-						</SliderColumn>
-						<SliderColumn>
-							<Widget label="Energy">
-								<Slider
-									value={filters.energy}
-									onChange={x => onFilterUpdate('energy', x)}
-								/>
-							</Widget>
-						</SliderColumn>
-						<SliderColumn>
-							<Widget label="Barkness">
-								<Slider
-									value={filters.barkness}
-									onChange={x => onFilterUpdate('barkness', x)}
-								/>
-							</Widget>
-						</SliderColumn>
-					</Flex>
-				</Card>
-				<Card>
-					<Flex m={-2} flexWrap="wrap">
-						{dogs.map(dog => (
-							<Box key={dog.id} p={2}>
-								<DogPhoto id={dog.id} breed={dog.name} />
-							</Box>
-						))}
-					</Flex>
-				</Card>
+				<DogFinder />
 			</Box>
 			<Box mb={4}>
 				<Heading size="gamma" is="h2" mb={2}>
