+++ src/components/DogFinder.md
@@ -1,0 +1,7 @@
+```jsx
+<DogFinder
+  dogs={[{id: 'saluki', name: 'Saluki'}, {id: 'afghan-hound', name: 'Afghan Hound'}, {id: 'english-setter', name: 'English Setter'}]}
+  filters={{ size: 3, energy: 3, barkness: 3}}
+  onFilterUpdate={() => {}}
+/>
+```
