+++ src/components/Dogs.md
@@ -1,0 +1,5 @@
+```jsx
+<Dogs
+  dogs={[{id: 'saluki', name: 'Saluki'}, {id: 'afghan-hound', name: 'Afghan Hound'}, {id: 'english-setter', name: 'English Setter'}]}
+/>
+```
